Fun with Images
I decided to see if I could make it so when I clicked on the images of the lake, I could see them larger. Ended up creating a page for each image.

I went through all the html pages and removed the extra slashes that while they validated, it said they were not necessary.

All pages validated, and getting ready to submit assignment.